% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_mean_cq}
\alias{calculate_mean_cq}
\title{Calculate the mean Cq value for each sample-target pair}
\usage{
calculate_mean_cq(data)
}
\arguments{
\item{data}{A validated data frame of all samples, targets, and wells. Must
be processed by \code{\link{highlight_outliers}} and
\code{\link{exclude_invalid_wells}}.}
}
\value{
A data frame with the mean Cq value for each sample-target pair.
}
\description{
This function calculates the mean Cq value for each sample-target
             pair after excluding invalid wells.
}
