% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_relative_expression}
\alias{calculate_relative_expression}
\title{Calculate relative expression}
\usage{
calculate_relative_expression(data, refs)
}
\arguments{
\item{data}{A validated data frame of all samples, targets, and wells. Must
be processed by \code{\link{highlight_outliers}} and
\code{\link{exclude_invalid_wells}}.}

\item{refs}{One or more targets that serve as the housekeeping gene(s)}
}
\value{
A list containing the raw data and the relative expression of each
        sample compared to the reference sample for each target.
}
\description{
This function calculates the relative expression of each sample
             compared to a reference sample for each target.
}
