% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{detect_outlier_well}
\alias{detect_outlier_well}
\title{Outlier detection in triplicate data}
\usage{
detect_outlier_well(cq_data)
}
\arguments{
\item{cq_data}{A vector of length 3 containing the Cq values for a given 
sample-target pair. NA values should be included.}
}
\value{
A logical vector indicating whether each well is an outlier. An NA
        value indicates that the well is missing and is always considered an
        outlier.
}
\description{
This function identifies outlier wells in triplicate data.
             If 3 wells are present:
               - Find the two closest values.
               - Mark the third one as an outlier if it differs more than 1 cycle.
               - If all three are different, mark all three.
             If 2 wells are present:
               - If they differ by more than 1, highlight both.
             If 1 well is present:
               - Highlight it.
}
