% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{exclude_invalid_wells}
\alias{exclude_invalid_wells}
\title{Exclude invalid wells}
\usage{
exclude_invalid_wells(data, exclude_wells)
}
\arguments{
\item{data}{A validated data frame of all samples, targets, and wells. Must
be processed by \code{\link{highlight_outliers}}.}

\item{exclude_wells}{A manually curated vector of wells to exclude. These
should be wells that are manually determined to be
invalid based on results from \code{\link{highlight_outliers}}
and \code{\link{get_invalid_wells}}. Only outlier wells
are necessary to include in this vector. NA wells will
be automatically excluded.}
}
\value{
A data frame with invalid wells marked to be excluded.
}
\description{
This function excludes invalid wells from the data.
}
