% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_invalid_wells}
\alias{get_invalid_wells}
\title{Get invalid wells}
\usage{
get_invalid_wells(data)
}
\arguments{
\item{data}{A validated data frame of all samples, targets, and wells. Must
be processed by \code{\link{highlight_outliers}}.}
}
\value{
A list of two vectors, one containing the wells with missing values
        and the other containing the wells with outlier values.
}
\description{
This function returns list of two vectors, one containing the
             wells with missing values and the other containing the wells
             with outlier values.
}
