% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{highlight_outliers}
\alias{highlight_outliers}
\title{Highlight outlier wells in triplicate data}
\usage{
highlight_outliers(data)
}
\arguments{
\item{data}{A validated data frame of all samples, targets, and wells.}
}
\value{
A data frame with an additional column indicating whether each well
        is an outlier.
}
\description{
This function highlights outlier wells in triplicate data.
}
\seealso{
\code{\link{detect_outlier_well}}
}
