% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data_import.R
\name{import_and_validate}
\alias{import_and_validate}
\title{Validate data import}
\usage{
import_and_validate(data)
}
\arguments{
\item{data}{Data to validate, a data frame or a path to a data file. The file
must be a standard file for qPCR data. Each row must be a well.
Columns should include the minimum required columns for qPCR data,
including: Well, Target, Sample, Cq, and Group.}
}
\value{
A data frame with the required columns for qPCR data. The column names
        are standardized to Well, Target, Sample, and Cq.
}
\description{
Validate data import
}
