% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_expr}
\alias{plot_expr}
\title{Plot the results of the qPCR data}
\usage{
plot_expr(results)
}
\arguments{
\item{results}{A list of dataframes containing the results of the qPCR data}
}
\value{
A list of ggplot2 objects
}
\description{
Plot the results of the qPCR data
}
